backend/requirements.txt
fastapi==0.88.0
jose==1.0.0
passlib==1.7.4
pydantic==1.10.2
python-dotenv==0.21.0
python_bcrypt==0.3.2
python_jose==3.3.0
SQLAlchemy==1.4.41
psycopg2==2.9.5
uvicorn==0.20.0
